--- community-puzzles/game-of-life/.gitignore
@@ -1,19 +1,0 @@
-# Created by .ignore support plugin (hsz.mobi)
-### Java template
-*.class
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-# IntelliJ IDEA files
-*.iml
-.idea/
-out/
